requirements.txt
# requirements.txt
psycopg2-binary==2.9.6
streamlit==1.21.0
pandas==1.3.5